+++ resources/inference_service.py
@@ -1,6 +1,6 @@
 from ocp_resources.resource import NamespacedResource
 
-from utils.constants import KSERVE_API_GROUP
+from utilities.constants import KSERVE_API_GROUP
 
 
 # TODO: Move this to openshift-python-wrapper once we are confident
