+++ resources/trustyai_service.py
@@ -1,6 +1,6 @@
 from ocp_resources.resource import NamespacedResource
 
-from utils.constants import TRUSTYAI_API_GROUP, TRUSTYAI_SERVICE_IMAGE
+from utilities.constants import TRUSTYAI_API_GROUP, TRUSTYAI_SERVICE_IMAGE
 
 
 # TODO: Move this to openshift-python-wrapper once we are confident
